Move commands to a commands package and improve some javadocs and names
