+++ src/test/java/ParserTest.java
@@ -1,13 +1,14 @@
-import duke.CommandParseException;
-import duke.Parser;
-import duke.UnmarkCommand;
-import org.junit.jupiter.api.Test;
-
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
 
+import org.junit.jupiter.api.Test;
+
+import duke.CommandParseException;
+import duke.Parser;
+import duke.commands.UnmarkCommand;
+
 public class ParserTest {
-    Parser parser = new Parser();
+    private static Parser parser = new Parser();
 
     @Test
     public void unmarkTask_invalidNumber_exceptionThrown() {
